+++ src/s_util.cpp
@@ -26,7 +26,7 @@ extern RxNode *CompletionFilter;
 static int copyfile(char const* f1, char const* f2) { // from F1 to F2
     void *buffer;
     int fd1, fd2;
-    int rd;
+    ssize_t rd;
 
     buffer = FileBuffer;
